LNS: Fix register_height uint16_t truncation (#1085)

+++ src/cryptonote_core/loki_name_system.cpp
@@ -131,7 +131,7 @@ static mapping_record sql_get_mapping_from_statement(sqlite3_stmt *statement)
     return result;
 
   result.type            = static_cast<mapping_type>(type_int);
-  result.register_height = static_cast<uint16_t>(sqlite3_column_int(statement, static_cast<int>(mapping_record_column::register_height)));
+  result.register_height = static_cast<uint64_t>(sqlite3_column_int(statement, static_cast<int>(mapping_record_column::register_height)));
   result.owner_id        = sqlite3_column_int(statement, static_cast<int>(mapping_record_column::owner_id));
   result.backup_owner_id = sqlite3_column_int(statement, static_cast<int>(mapping_record_column::backup_owner_id));
